ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.8",
    "boto3>=1.42.14",
    "fastapi>=0.121.3",
    "langchain>=1.0.8",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "loguru>=0.7.3",
    "prisma>=0.15.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "python-socketio>=5.15.0",
    "uvicorn>=0.38.0",
]
